# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"Player": {
			"properties": {
				"_id": {
					"type": "string"
				},
				"prename": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"number": {
					"type": "number",
					"format": "double"
				},
				"birthday": {
					"type": "string",
					"x-nullable": true
				},
				"coach": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"x-nullable": true
				},
				"team": {
					"type": "array",
					"items": {
						"type": "string"
					},
					"x-nullable": true
				}
			},
			"required": [
				"_id",
				"prename",
				"name",
				"number"
			],
			"type": "object"
		},
...
		"Goals": {
			"properties": {
				"shot": {
					"type": "number",
					"format": "double"
				},
				"got": {
					"type": "number",
					"format": "double"
				},
				"difference": {
					"type": "number",
					"format": "double",
					"x-nullable": true
				}
			},
			"required": [
				"shot",
				"got"
			],
			"type": "object"
		},
		"Statistic": {
			"properties": {
				"gameWon": {
					"type": "object"
				},
				"goalsShot": {
					"type": "object"
				},
				"goalsGot": {
					"type": "object"
				},
				"highestWin": {
					"type": "object"
				},
				"highestLose": {
					"type": "object"
				}
			},
			"required": [
				"gameWon",
				"goalsShot",
				"goalsGot",
				"highestWin",
				"highestLose"
			],
			"type": "object"
		},
		"Score": {
			"properties": {
				"id": {
					"type": "string"
				},
				"rank": {
					"type": "number",
					"format": "double"
				},
				"name": {
					"type": "string"
				},
				"live": {
					"type": "boolean"
				},
				"games": {
					"type": "object"
				},
				"goals": {
					"$ref": "#/definitions/Goals"
				},
				"points": {
					"type": "object"
				},
				"statistics": {
					"$ref": "#/definitions/Statistic",
					"x-nullable": true
				}
			},
			"required": [
				"id",
				"rank",
				"name",
				"live",
				"games",
				"goals",
				"points"
			],
			"type": "object"
		},
		"Gymnasium": {
			"properties": {
				"id": {
					"type": "string"
				},
				"number": {
					"type": "string"
				},
				"name": {
					"type": "string",
					"x-nullable": true
				},
				"street": {
					"type": "string",
					"x-nullable": true
				},
				"postal": {
					"type": "string",
					"x-nullable": true
				},
				"city": {
					"type": "string",
					"x-nullable": true
				}
			},
			"required": [
				"id",
				"number"
			],
			"type": "object"
		},
		"Teams": {
			"properties": {
				"home": {
					"type": "number",
					"format": "double"
				},
				"guest": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"home",
				"guest"
			],
			"type": "object"
		},
		"gameGoals": {
			"properties": {
				"end": {
					"$ref": "#/definitions/Teams"
				},
				"halfTime": {
					"$ref": "#/definitions/Teams"
				}
			},
			"required": [
				"end",
				"halfTime"
			],
			"type": "object"
		},
		"Game": {
			"properties": {
				"id": {
					"type": "string"
				},
				"live": {
					"type": "boolean"
				},
				"date": {
					"type": "string",
					"format": "date-time"
				},
				"gymnasium": {
					"$ref": "#/definitions/Gymnasium"
				},
				"team": {
					"type": "object"
				},
				"goals": {
					"$ref": "#/definitions/gameGoals"
				},
				"points": {
					"$ref": "#/definitions/Teams"
				},
				"referees": {
					"type": "string"
				},
				"comment": {
					"type": "string",
					"x-nullable": true
				},
				"sortText": {
					"type": "string",
					"x-nullable": true
				},
				"token": {
					"type": "string",
					"x-nullable": true
				},
				"appId": {
					"type": "string",
					"x-nullable": true
				},
				"report": {
					"type": "string",
					"x-nullable": true
				}
			},
			"required": [
				"id",
				"live",
				"date",
				"gymnasium",
				"team",
				"goals",
				"points",
				"referees"
			],
			"type": "object"
		},
		"Ligue": {
			"properties": {
				"id": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"shortName": {
					"type": "string"
				},
				"comment": {
					"type": "string",
					"x-nullable": true
				},
				"headline1": {
					"type": "string",
					"x-nullable": true
				},
				"headline2": {
					"type": "string",
					"x-nullable": true
				},
				"actualized": {
					"type": "string",
					"x-nullable": true
				},
				"scores": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Score"
					},
					"x-nullable": true
				},
				"games": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Game"
					}
				}
			},
			"required": [
				"id",
				"name",
				"shortName",
				"games"
			],
			"type": "object"
		},
		"Club": {
			"properties": {
				"id": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"shortName": {
					"type": "string"
				},
				"headline1": {
					"type": "string",
					"x-nullable": true
				},
				"headline2": {
					"type": "string",
					"x-nullable": true
				},
				"actualized": {
					"type": "string",
					"x-nullable": true
				},
				"ligues": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Ligue"
					},
					"x-nullable": true
				}
			},
			"required": [
				"id",
				"name",
				"shortName"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "wh_middleware",
		"version": "0.0.9",
		"description": "TSV Willsbach Handball API's",
		"license": {
			"name": "MIT"
		}
	},
	"paths": {
		"/players": {
			"post": {
				"operationId": "AddNewPlayer",
				"produces": [
					"application/json"
				],
				"responses": {
					"201": {
						"description": "Player created"
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "requstBody",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Player"
						}
					}
				]
			},
			"get": {
				"operationId": "GetPlayers",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": []
			}
		},
		"/players/{id}": {
			"get": {
				"operationId": "GetSinglePlayer",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"delete": {
				"operationId": "DeletePlayer",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "UpdatePlayer",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Players"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "requstBody",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Player"
						}
					}
				]
			}
		},
...
		"/hvw/club/{id}": {
			"get": {
				"operationId": "GetClub",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Club"
						}
					}
				},
				"tags": [
					"HVW"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/hvw/games": {
			"get": {
				"operationId": "GetGames",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Game"
							}
						}
					},
					"400": {
						"description": "Bad request"
					}
				},
				"tags": [
					"HVW"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/hvw/ligue": {
			"get": {
				"operationId": "GetLigue",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Ligue"
						}
					}
				},
				"tags": [
					"HVW"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {},
...
}